requirements.txt
# requirements.txt
Flask==2.3.3
pymongo==4.5.0
dnspython==2.4.2
python-dotenv==1.0.0
flask-cors==4.0.0
spacy==3.7.2
fr-core-news-sm==3.7.0
scikit-learn==1.3.0
pandas==2.1.1
numpy==1.25.2
PyJWT==2.8.0
certifi==2023.11.17